+++ models/microsoft/graph/drive_item_versionable.go
@@ -9,7 +9,7 @@ type DriveItemVersionable interface {
     BaseItemVersionable
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
     GetContent()([]byte)
-    GetSize()(*int64)
+    GetSize()(*int32)
     SetContent(value []byte)()
-    SetSize(value *int64)()
+    SetSize(value *int32)()
 }
