+++ models/microsoft/graph/integer_range.go
@@ -9,9 +9,9 @@ type IntegerRange struct {
     // Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well.
     additionalData map[string]interface{};
     // The inclusive upper bound of the integer range.
-    end *int64;
+    end *int32;
     // The inclusive lower bound of the integer range.
-    start *int64;
+    start *int32;
 }
 // NewIntegerRange instantiates a new integerRange and sets the default values.
 func NewIntegerRange()(*IntegerRange) {
@@ -33,7 +33,7 @@ func (m *IntegerRange) GetAdditionalData()(map[string]interface{}) {
     }
 }
 // GetEnd gets the end property value. The inclusive upper bound of the integer range.
-func (m *IntegerRange) GetEnd()(*int64) {
+func (m *IntegerRange) GetEnd()(*int32) {
     if m == nil {
         return nil
     } else {
@@ -44,7 +44,7 @@ func (m *IntegerRange) GetEnd()(*int64) {
 func (m *IntegerRange) GetFieldDeserializers()(map[string]func(interface{}, i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode)(error)) {
     res := make(map[string]func(interface{}, i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode)(error))
     res["end"] = func (o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-        val, err := n.GetInt64Value()
+        val, err := n.GetInt32Value()
         if err != nil {
             return err
         }
@@ -54,7 +54,7 @@ func (m *IntegerRange) GetFieldDeserializers()(map[string]func(interface{}, i04e
         return nil
     }
     res["start"] = func (o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-        val, err := n.GetInt64Value()
+        val, err := n.GetInt32Value()
         if err != nil {
             return err
         }
@@ -66,7 +66,7 @@ func (m *IntegerRange) GetFieldDeserializers()(map[string]func(interface{}, i04e
     return res
 }
 // GetStart gets the start property value. The inclusive lower bound of the integer range.
-func (m *IntegerRange) GetStart()(*int64) {
+func (m *IntegerRange) GetStart()(*int32) {
     if m == nil {
         return nil
     } else {
@@ -79,13 +79,13 @@ func (m *IntegerRange) IsNil()(bool) {
 // Serialize serializes information the current object
 func (m *IntegerRange) Serialize(writer i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.SerializationWriter)(error) {
     {
-        err := writer.WriteInt64Value("end", m.GetEnd())
+        err := writer.WriteInt32Value("end", m.GetEnd())
         if err != nil {
             return err
         }
     }
     {
-        err := writer.WriteInt64Value("start", m.GetStart())
+        err := writer.WriteInt32Value("start", m.GetStart())
         if err != nil {
             return err
         }
@@ -105,13 +105,13 @@ func (m *IntegerRange) SetAdditionalData(value map[string]interface{})() {
     }
 }
 // SetEnd sets the end property value. The inclusive upper bound of the integer range.
-func (m *IntegerRange) SetEnd(value *int64)() {
+func (m *IntegerRange) SetEnd(value *int32)() {
     if m != nil {
         m.end = value
     }
 }
 // SetStart sets the start property value. The inclusive lower bound of the integer range.
-func (m *IntegerRange) SetStart(value *int64)() {
+func (m *IntegerRange) SetStart(value *int32)() {
     if m != nil {
         m.start = value
     }
