+++ models/microsoft/graph/integer_rangeable.go
@@ -8,8 +8,8 @@ import (
 type IntegerRangeable interface {
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.AdditionalDataHolder
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
-    GetEnd()(*int64)
-    GetStart()(*int64)
-    SetEnd(value *int64)()
-    SetStart(value *int64)()
+    GetEnd()(*int32)
+    GetStart()(*int32)
+    SetEnd(value *int32)()
+    SetStart(value *int32)()
 }
