+++ models/microsoft/graph/print_documentable.go
@@ -10,8 +10,8 @@ type PrintDocumentable interface {
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
     GetContentType()(*string)
     GetDisplayName()(*string)
-    GetSize()(*int64)
+    GetSize()(*int32)
     SetContentType(value *string)()
     SetDisplayName(value *string)()
-    SetSize(value *int64)()
+    SetSize(value *int32)()
 }
