+++ models/microsoft/graph/tone_infoable.go
@@ -8,8 +8,8 @@ import (
 type ToneInfoable interface {
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.AdditionalDataHolder
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
-    GetSequenceId()(*int64)
+    GetSequenceId()(*int32)
     GetTone()(*Tone)
-    SetSequenceId(value *int64)()
+    SetSequenceId(value *int32)()
     SetTone(value *Tone)()
 }
